MetadataServer/requirements.txt
# This is the master file; see deploy_requirements.txt for production.
# If necessary we could have a core set of requirements, and source
# that in for differences between prod/testing (eg, psycopg2 only in
# prod, and django-extensions only for testing)

Django==1.8.13
djangorestframework==3.3.3
django-extensions==1.6.7
pytz==2016.4
lxml==3.6.0
-e git://github.com/condense/django-allauth.git@fix-email-required-widget#egg=django-allauth
django-uuidfield==0.5.0
xmlunittest==0.3.2
django-bootstrap3==7.0.1
jsonfield==1.0.3
djangorestframework-jsonp==1.0.2
djangorestframework-xml==1.3.0
Pillow==3.3.0
django-imagekit==3.3
django-fsm==2.4.0
django-fsm-admin==1.2.3
requests==2.10.0
django-filter==0.13.0
django-treebeard==4.0.1
